pyproject.toml
[tool.poetry]
name = "olapy"
version = "0.8.2"
authors = ["Abilian SAS"]
description = "OlaPy, an experimental OLAP engine based on Pandas"
classifiers = [
  "Programming Language :: Python",
  "Development Status :: 3 - Alpha",
  "Natural Language :: English",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9"
]
homepage = "https://github.com/abilian/olapy"
readme = "README.rst"

packages = [{ include = "olapy", from = "src" }]


[tool.poetry.scripts]
olapy = 'olapy.__main__:cli'
etl = 'olapy.etl.etl:run_etl'


[tool.poetry.dependencies]
python = ">=3.8,<4"

pandas = "*"
attrs = "*"
click = "*"
lxml = "*"
pyyaml = ">=4.2b1"
regex = "*"
spyne = "^2.13"
sqlalchemy = "*"
typing = "*"
xmlwitch = "*"

# Extra (see below)
awesome-slugify = {optional = true, version = "*"}
bonobo = {optional = true, version = "*"}
bonobo-sqlalchemy = {version = "<0.6.1", optional = true}
pyspark = {version = "<3", optional = true}
python-dotenv = {optional = true, version = "*"}
whistle = {version = "<1.0.1", optional = true}


[tool.poetry.extras]
etl = ["awesome-slugify", "bonobo", "bonobo-sqlalchemy", "python-dotenv", "whistle"]
spark = ["pyspark"]


[tool.poetry.dev-dependencies]
# Tools / dev / testing
tox = "*"
pytest = "*"
pytest-cov = "*"
PrettyTable = "*"
gprof2dot = "*"
ipdb = "*"
py-cpuinfo = "*"

#databases_connectors
#pyodbc #need some packages to be installed (sudo apt-get install unixodbc-dev)
#cx_oracle
#mysql-python # don't work in py3
#pymysql

# linting
mypy = "*"
flake8 = "*"
flake8-bugbear = "*"
flake8-comprehensions = "*"
flake8-mutable = "*"
flake8-pytest = "*"
flake8-super-call = "*"
flake8-tidy-imports = "*"
mastool = "*"
mccabe = "*"
pylint = "*"
pre-commit = "*"

# Formatting
isort = "^5"
black = "*"

# FIXME
# xmla (for testing - Python 2 only due to dependency on suds)
# xmla = { version = "*", python = "^2.7" }

# Doc
sphinx-click = "*"
sphinx-rtd-theme = "*"
types-PyYAML = "^6.0.1"
types-six = "^1.16.3"
devtools = "^0.8.0"
